Revert "Calculate storage limit"
