+++ app/dashboard/[slug]/component.tsx
@@ -13,15 +13,8 @@ import {
   ConnectNotificationBanner,
 } from "@stripe/react-connect-js";
 import { useRouter } from "next/navigation";
-import General from "@/components/general";
 
-export default function Component({
-  slug,
-  storageUsed,
-}: {
-  slug: string;
-  storageUsed: bigint;
-}) {
+export default function Component({ slug }: { slug: string }) {
   const router = useRouter();
   const component = useMemo(() => {
     switch (slug) {
@@ -54,12 +47,10 @@ export default function Component({
         return <ConnectPayouts />;
       case "notifications":
         return <ConnectNotificationBanner />;
-      case "general":
-        return <General storageUsed={storageUsed} />;
       default:
         return <div>Not found</div>;
     }
-  }, [slug, router, storageUsed]);
+  }, [slug, router]);
 
   return (
     <div className="w-full h-fit md:py-24 pb-40 flex items-center justify-center relative">
