+++ app/upload/page.tsx
@@ -1,6 +1,5 @@
 import UploadPage from "@/components/upload/multi-step";
-import { STORAGE_LIMIT_50_GB_IN_BYTES } from "@/lib/consts";
-import { doesUserHaveStripeAccount, getUserUsedStorage } from "@/lib/db/mod";
+import { doesUserHaveStripeAccount } from "@/lib/db/mod";
 import { hasRequirementsDue } from "@/lib/stripe";
 import { currentUser } from "@clerk/nextjs/server";
 import { notFound, redirect } from "next/navigation";
@@ -20,11 +19,6 @@ export default async function UploadPageRoot() {
   const hasRequirements = await hasRequirementsDue();
   if (hasRequirements) {
     redirect("dashboard/onboarding");
-  }
-
-  const storageUsed = await getUserUsedStorage(userData.username);
-  if (!storageUsed || storageUsed >= STORAGE_LIMIT_50_GB_IN_BYTES) {
-    redirect("/dashboard/onboarding");
   }
 
   return <UploadPage userName={userData.username} />;
