+++ cfg.ts
@@ -23,5 +23,3 @@ function getEnvVar(name: string) {
   }
   return value;
 }
-
-
