+++ components/app-sidebar.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import Link from "next/link";
 import { Badge } from "@/components/ui/badge";
 import {
   Sidebar,
@@ -17,18 +16,14 @@ import {
 } from "@/components/ui/sidebar";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { cn } from "@/lib/utils";
+import Link from "next/link";
 import { useParams } from "next/navigation";
 import { Suspense } from "react";
 
 const items = [
   {
-    title: "General",
-    subItems: [
-      { title: "General", url: "/dashboard/general", slug: "general" },
-    ],
-  },
-  {
     title: "Stripe",
+    url: "/dashboard",
     subItems: [
       {
         title: "Payments",
@@ -95,9 +90,11 @@ export function AppSidebar({ children }: AppSidebarProps) {
               {items.map((item) => (
                 <SidebarMenuItem key={item.title}>
                   <SidebarMenuButton asChild>
-                    <span className="font-bold text-neutral-100">
-                      {item.title}
-                    </span>
+                    <Link href={item.url}>
+                      <span className="font-bold text-neutral-100">
+                        {item.title}
+                      </span>
+                    </Link>
                   </SidebarMenuButton>
                   <SidebarMenuSub>
                     {item.subItems.map((subItem) => (
