+++ components/upload/info-and-controls.tsx
@@ -1,11 +1,10 @@
 import NumberFlow from "@number-flow/react";
-import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
-import { UseFormReturn } from "react-hook-form";
 import { UploadFormSchema } from "@/components/upload-form-schema";
+import { cn } from "@/lib/utils";
 import { ArrowLeft, Loader, ArrowRight } from "lucide-react";
-
-import type { Dispatch, SetStateAction } from "react";
+import { Dispatch, SetStateAction } from "react";
+import { UseFormReturn } from "react-hook-form";
 
 export default function StepInfoAndControls({
   stepIndex,
@@ -35,12 +34,6 @@ export default function StepInfoAndControls({
     setStepIndex(stepIndex + 1);
   }
 
-  function handlePreviousClick() {
-    if (isFirstStep) return;
-    setDirection(-1);
-    setStepIndex(stepIndex - 1);
-  }
-
   return (
     <div className="flex flex-col items-start justify-start w-full">
       <div className="w-full flex items-center justify-between">
@@ -67,7 +60,13 @@ export default function StepInfoAndControls({
       <div className="w-full flex items-center justify-between pb-10 mt-1">
         <Button
           variant="secondary"
-          onClick={handlePreviousClick}
+          onClick={() => {
+            if (isFirstStep) {
+              return;
+            }
+            setStepIndex(stepIndex - 1);
+            setDirection(-1);
+          }}
           disabled={isFirstStep || isLoading || isSuccess}
           className="active:scale-90 w-10"
         >
