+++ components/upload/steps/four.tsx
@@ -1,6 +1,3 @@
-import { Input } from "@/components/ui/input";
-import { AnimatePresence, motion } from "framer-motion";
-
 import {
   FormField,
   FormItem,
@@ -8,7 +5,8 @@ import {
   FormControl,
   FormMessage,
 } from "@/components/ui/form";
-
+import { Input } from "@/components/ui/input";
+import { AnimatePresence, motion } from "framer-motion";
 import type { UploadFormSchema } from "@/components/upload-form-schema";
 import type { UseFormReturn } from "react-hook-form";
