+++ components/upload/steps/six.tsx
@@ -2,7 +2,6 @@ import Link from "next/link";
 import { Button } from "@/components/ui/button";
 import { AnimatePresence, motion } from "framer-motion";
 import { Loader, Check } from "lucide-react";
-
 import type { UseFormReturn } from "react-hook-form";
 import type { UploadFormSchema } from "@/components/upload-form-schema";
