+++ lib/aws/client.ts
@@ -1,9 +1,5 @@
 import { S3Client } from "@aws-sdk/client-s3";
-import {
-  AWS_REGION,
-  AWS_ACCESS_KEY_ID,
-  AWS_SECRET_ACCESS_KEY,
-} from "@/lib/cfg";
+import { AWS_REGION, AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY } from "@/cfg";
 import { isDev, isTest } from "@/lib/utils";
 
 const clientCfg =
