+++ lib/aws/mod.ts
@@ -5,7 +5,7 @@ import "server-only";
 import log from "@/lib/log";
 import s3 from "@/lib/aws/client";
 import crypto from "node:crypto";
-import { AWS_PRIVATE_BUCKET_NAME, AWS_PUBLIC_BUCKET_NAME } from "@/lib/cfg";
+import { AWS_PRIVATE_BUCKET_NAME, AWS_PUBLIC_BUCKET_NAME } from "@/cfg";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { isDev } from "@/lib/utils";
 import {
