+++ lib/db/cfg/schema.prisma
@@ -1,5 +1,6 @@
+
 generator client {
-  provider      = "prisma-client-js"
+  provider = "prisma-client-js"
   binaryTargets = ["native", "rhel-openssl-3.0.x"]
 }
 
@@ -17,7 +18,6 @@ model User {
   imgUrl      String
   stripeId    String?
   name        String?
-  storageUsed BigInt        @default(0)
   samplePacks SamplePack[]
   createdAt   DateTime      @default(now())
   Transaction Transaction[]
@@ -37,7 +37,6 @@ model SamplePack {
   creator           User          @relation(fields: [creatorId], references: [id], onDelete: Cascade)
   creatorId         Int
   createdAt         DateTime      @default(now())
-  totalSize         BigInt           @default(0)
   Transaction       Transaction[]
 
   @@unique([creatorId, name])
@@ -64,7 +63,7 @@ model Transaction {
 }
 
 model InterestedUser {
-  id        Int      @id @default(autoincrement())
-  email     String   @unique
-  createdAt DateTime @default(now())
+  id           Int        @id @default(autoincrement())
+  email        String     @unique
+  createdAt    DateTime   @default(now())
 }
